tests/wpt/tests/css/css-overflow/scroll-marker-focus-within.html

<!doctype html>
<meta charset="utf-8">
<title>CSS Overflow Test: focused ::scroll-marker sets :focus-within on its ::scroll-marker-group, scroller and all the ancestors</title>
<link rel="help" href="https://drafts.csswg.org/css-overflow-5/#focus-pseudoclasses">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<style>
  body {
    margin: 0;
  }

  #scroller {
    background: red;
    width: 600px;
    height: 300px;
    overflow: auto;
    scroll-marker-group: before;
    white-space: nowrap;
  }

  #scroller:focus-within {
    background: green;
  }

  #scroller div {
    background: green;
    display: inline-block;
    width: 600px;
    height: 270px;
  }

  #scroller div:focus-within {
    background: red;
  }

  #scroller::scroll-marker-group {
    background: red;
    display: flex;
    height: 20px;
    width: 40px;
  }

  #scroller::scroll-marker-group:focus-within {
    background: green;
  }

  #scroller div::scroll-marker {
    content: "";
    width: 20px;
    height: 20px;
    background-color: red;
    display: inline-block;
  }

  #scroller div::scroll-marker:focus-within {
    background: green;
  }
</style>
<div id="scroller">
  <div id="target"></div>
  <div></div>
</div>
<script>
  promise_test(async t => {
    let actions_promise = new test_driver.Actions()
      .pointerMove(15, 15)
      .pointerDown()
      .pointerUp()
      .send();
    await actions_promise;
    assert_equals(getComputedStyle(scroller, "::scroll-marker-group").backgroundColor, "rgb(0, 128, 0)", "focused ::scroll-marker sets :focus-within on its ::scroll-marker-group");
    assert_equals(getComputedStyle(scroller).backgroundColor, "rgb(0, 128, 0)", "focused ::scroll-marker sets :focus-within on its scroller");
    assert_equals(getComputedStyle(target, "::scroll-marker").backgroundColor, "rgb(0, 128, 0)", "focused ::scroll-marker sets :focus-within on itself");
    assert_equals(getComputedStyle(target).backgroundColor, "rgb(0, 128, 0)", "focused ::scroll-marker doesn't set :focus-within on its originating element");
    actions_promise = new test_driver.Actions()
      .pointerMove(500, 300)
      .pointerDown()
      .pointerUp()
      .send();
    await actions_promise;
    assert_equals(getComputedStyle(scroller, "::scroll-marker-group").backgroundColor, "rgb(255, 0, 0)", "focused ::scroll-marker resets :focus-within on its ::scroll-marker-group");
    assert_equals(getComputedStyle(scroller).backgroundColor, "rgb(255, 0, 0)", "focused ::scroll-marker resets :focus-within on its scroller");
    assert_equals(getComputedStyle(target, "::scroll-marker").backgroundColor, "rgb(255, 0, 0)", "focused ::scroll-marker resets :focus-within on itself");
  });
</script>
