third_party/blink/web_tests/external/wpt/css/css-overflow/scroll-marker-focus-within-expected.txt

This is a testharness.js-based test.
[FAIL] CSS Overflow Test: focused ::scroll-marker sets :focus-within on its ::scroll-marker-group, scroller and all the ancestors
  assert_equals: focused ::scroll-marker sets :focus-within on its ::scroll-marker-group expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
Harness: the test ran to completion.
